// frontend/src/types/types.ts
export interface ILoginResponse {
  auth: IAuth;
  user: IUser;
}

export interface IRegisterResponse {
  data: IUser;
}

export interface ICheckSession {
  auth: IAuth;
  user: IUser;
}

export interface IAuth {
  id: string;
  email: string;
  username: string;
}
// ...
export interface IUser {
  id: string;
  username: string;
  fullName: string;
  email: string;
  profilePicture?: string | Record<string, any>;
  isEmailValidated: boolean;
  [prop: string]: any;
}